Solve library with toggle readMark

+++ app.js
@@ -13,7 +13,12 @@ function addBookToLibrary() {
   const author = document.getElementById("author").value;
   const title = document.getElementById("title").value;
   const pages = document.getElementById("pages").value;
-  const read = document.getElementById("readYes").value;
+  let read = document.getElementById("read").checked;
+  if (read) {
+    read = "yes";
+  } else {
+    read = "no";
+  }
   const add = new Book(author, title, pages, read);
   myLibrary.push(add);
 }
@@ -23,7 +28,6 @@ function printLibrary() {
   container.innerHTML = "";
   for (let i = 0; i < myLibrary.length; i++) {
     let card = document.createElement('div');
-    card.id = i;
     card.innerHTML = `
     <div>
     <div class="card">
@@ -33,6 +37,7 @@ function printLibrary() {
         <p class="card-text"># Pages: ${myLibrary[i].pages}</p>
         <p class="card-text">Read: ${myLibrary[i].read}</p>
         <a href="#" class="btn btn-warning delete" data-index="${i}">Delete</a>
+        <a href="#" class="btn btn-warning toggle" data-toggle="${i}">Read</a>
       </div>
     </div>
     </div>`
@@ -43,18 +48,35 @@ function printLibrary() {
 // remove instance from myLibrary array
 function deleteBookFromLibrary(erase) {
   myLibrary.splice(erase, 1);
+  if (myLibrary.length < 1) {
+    myLibrary = [];
+  }
+}
+// Toggle Read instance from myLibrary array
+function readBookFromLibrary(toggle) {
+  if (myLibrary[toggle].read === "yes") {
+    myLibrary[toggle].read = "no";
+  } else {
+    myLibrary[toggle].read = "yes";
+  }
 }
 // When submit is clicked, prevent default of submit and run functions
-const submit = document.getElementById("submit");
-submit.addEventListener("click", function(event) {
-  event.preventDefault();
+const form = document.querySelector("form");
+form.addEventListener("submit", function(event) {
   addBookToLibrary();
   printLibrary();
+  event.preventDefault();
 })
 // When delete is clicked, run functions
-const remove = document.getElementById("libraryShelf");
-remove.addEventListener("click", function(event) {
-  const erase = event.target.dataset.index;
-  deleteBookFromLibrary(erase);
-  printLibrary();
+const shelf = document.getElementById("libraryShelf");
+shelf.addEventListener("click", function(event) {
+  if (event.target.textContent === "Delete") {
+    const erase = event.target.dataset.index;
+    deleteBookFromLibrary(erase);
+    printLibrary();
+  } else {
+    const toggle = event.target.dataset.toggle;
+    readBookFromLibrary(toggle);
+    printLibrary();
+  }
 })
